Move CameraState and related classes to their own file
 Moves CameraState and VideoRecordingState out of the
 CameraSystem interface definition file into their own
 files.
